+++ examples/pfr.rs
@@ -60,7 +60,7 @@ fn main() -> ! {
 
 
     let clocks = hal::ClockRequirements::default()
-        .system_frequency(96.mhz())
+        .system_frequency(96.MHz())
         .configure(&mut anactrl, &mut pmc, &mut syscon)
         .unwrap();